So, this was supposed to be a simple extension...
...bind a few functions, let xsubpp do all the work, and so on...

Then I started thinking 'but I can make this EASIER', and suddenly
there's yards of cpp and typemaps and mess everywhere. <sigh>

It's important that the fluidsynth objects are destroyed in the right
order. If (say) a Synth is destroyed before a Driver which references it
we will get segfaults. So, to avoid Perl people having to deal with
that, I've made those objects which hold refs internally also hold
proper Perl refs on the equivalent Perl objects, so in principle nothing
should get destroyed out of order.

HOWEVER, if one of these objects gets as far as the final phase of
global destruction (clearing up cycles), all bets are off. (Perl doesn't
seem to make any effort to find individual cycles and break them, it
just destructs everything out of order.) There's really very little I
can do about this, except document Not To Do That, but I can at least
issue a warning if I see it's happened.

Also, check if construction failed, which it does sometimes.